remove custom futureTask, use standard CompletableFuture instead.
MasterFutureTask and SlaveFutureTask is not clear and error-prone.
When I fork this plugin and do some custom, I encounter a dead lock issue
because of the LOCK object in BranchInfo. These locks are useless since fetch & diff operation is fast.
So I replace with CompletableFuture which supports the same async effect and is clearer.